+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - You can now change the number of spaces to indent by in the preferences.
 - Images can now be put inside links on the condition that they are (a) the only element inside the link's description and (b) relative links won't work during preview.
 - You can now activate RMarkdown file support in the advanced preferences.
+- You can now tell Zettlr to count characters instead of words (e.g. for Chinese).
 
 ## Under the Hood